Changed the ints to doubles for all tests

+++ src/main/java/CertificateOfDeposit.java
@@ -1,6 +1,6 @@
 public class CertificateOfDeposit extends Account {
 
-    public CertificateOfDeposit(String identifier, double apr, int startAmount) {
+    public CertificateOfDeposit(String identifier, double apr, double startAmount) {
         super("CD", identifier, apr);
         this.setBalance(startAmount);
     }
